intergate with aria2

+++ src/hqr/util/SaveImgAndMagnet.java
@@ -47,14 +47,22 @@ public class SaveImgAndMagnet {
 		String []arr = imgUrl.split("/");
 		String fileName = arr[arr.length-1];
 		HttpGet get = new HttpGet(imgUrl);
-		try(
-			CloseableHttpResponse cl = httpclient.execute(get, httpClientContext);
-			FileOutputStream fos = new FileOutputStream((new File(folder+System.getProperty("file.separator")+fileName)));
-		){
-			cl.getEntity().writeTo(fos);
+		
+		File f = new File(folder+System.getProperty("file.separator")+fileName);
+		if(!f.exists()) {
+			try(
+					CloseableHttpResponse cl = httpclient.execute(get, httpClientContext);
+					FileOutputStream fos = new FileOutputStream((new File(folder+System.getProperty("file.separator")+fileName)));
+				){
+					cl.getEntity().writeTo(fos);
+					fos.flush();
+				}
+				catch (Exception e) {
+					e.printStackTrace();
+				}
 		}
-		catch (Exception e) {
-			e.printStackTrace();
+		else {
+			System.out.println(folder+System.getProperty("file.separator")+fileName+" exist, skip it");
 		}
 	}
